+++ app/graphql/mutations/create_project.rb
@@ -3,10 +3,10 @@
 module Mutations
   class CreateProject < BaseMutation
     class CreateProjectAddressAttributes < Types::BaseInputObject
-      argument :street, String, required: true
-      argument :street_no, String, required: true
-      argument :city, String, required: true
-      argument :zip, String, required: true
+      argument :street, String, required: false
+      argument :street_no, String, required: false
+      argument :city, String, required: false
+      argument :zip, String, required: false
     end
 
     class CreateProjectAddressBookAttributes < Types::BaseInputObject
@@ -15,6 +15,13 @@ module Mutations
       argument :type, String, required: true
       argument :name, String, required: false
       argument :address, CreateProjectAddressAttributes, as: :address_attributes, required: false
+    end
+
+    class CreateSiteAddressAttributes < Types::BaseInputObject
+      argument :apartments_count, Int, required: true
+      argument :name, String, required: false
+      argument :move_in_starts_on, GraphQL::Types::ISO8601DateTime, required: true
+      argument :address, CreateProjectAddressAttributes, required: false
     end
 
     class CreateProjectAttributes < Types::BaseInputObject
@@ -28,17 +35,19 @@ module Mutations
       argument :coordinate_north, Float, required: false
 
       argument :assignee_id, ID, required: false
+      argument :kam_assignee_id, ID, required: false
       argument :lot_number, String, required: false
       argument :buildings_count, Int, required: true
-      argument :apartments_count, Int, required: true
 
       argument :building_type, String, required: false
+      argument :description_on_other, String, required: false
       argument :construction_type, String, required: false
       argument :move_in_starts_on, GraphQL::Types::ISO8601DateTime, required: false
       argument :move_in_ends_on, GraphQL::Types::ISO8601DateTime, required: false
       argument :construction_starts_on, GraphQL::Types::ISO8601DateTime, required: false
 
       argument :address, CreateProjectAddressAttributes, required: false, as: :address_attributes
+      argument :site_address, [CreateSiteAddressAttributes], required: false
       argument :address_books, [CreateProjectAddressBookAttributes], required: true, as: :address_books_attributes
     end