+++ app/services/projects/building_updater.rb
@@ -5,7 +5,7 @@ module Projects
     include BuildingHelper
 
     def call
-      authorize! building.project, to: :update?
+      authorize! project, to: :update?
 
       with_tracking { building.update!(attributes) }
     end