+++ app/services/projects/transitions/callbacks.rb
@@ -81,6 +81,7 @@ module Projects
         default_label_group.label_list.delete_if { |label| AdminToolkit::PctValue.statuses.value?(label) }
 
         default_label_group.label_list << AdminToolkit::PctValue.statuses[pct_value.status]
+        update_prio_status(AdminToolkit::PctValue.statuses[pct_value.status])
         default_label_group.save!
       rescue StandardError => e
         raise(t('projects.transition.error_while_adding_label', error: e.message))
@@ -88,6 +89,10 @@ module Projects
 
       def update_exceeding_cost
         project.update!(exceeding_cost: ::AdminToolkit::CostThreshold.first.exceeding)
+      end
+
+      def update_prio_status(status)
+        project.update!(prio_status: status)
       end
     end
   end