+++ db/structure.sql
@@ -857,7 +857,13 @@ CREATE TABLE public.projects (
     cable_installations text[] DEFAULT '{}'::text[],
     priority_tac character varying,
     access_technology_tac character varying,
-    exceeding_cost numeric
+    exceeding_cost numeric,
+    site_address jsonb,
+    file_upload boolean DEFAULT false,
+    kam_assignee_id uuid,
+    confirmation_status character varying,
+    description_on_other character varying,
+    prio_status character varying
 );
 
 
@@ -896,13 +902,12 @@ CREATE TABLE public.projects_address_books (
 
 CREATE TABLE public.projects_buildings (
     id uuid DEFAULT public.gen_random_uuid() NOT NULL,
-    name character varying NOT NULL,
+    name character varying,
     external_id character varying,
     assignee_id uuid,
     project_id uuid NOT NULL,
     apartments_count integer DEFAULT 0 NOT NULL,
     move_in_starts_on date,
-    move_in_ends_on date,
     additional_details jsonb DEFAULT '{}'::jsonb,
     files_count integer DEFAULT 0 NOT NULL,
     tasks_count integer DEFAULT 0 NOT NULL,
@@ -1012,6 +1017,7 @@ CREATE MATERIALIZED VIEW public.projects_lists AS
     cardinality(projects.label_list) AS labels,
     addresses.city,
     addresses.zip,
+    projects.kam_assignee_id AS kam_assignee,
     concat(addresses.street, ' ', addresses.street_no, ', ', addresses.zip, ', ', addresses.city) AS address,
     COALESCE(NULLIF(concat(profiles.firstname, ' ', profiles.lastname), ' '::text), (projects.assignee_type)::text) AS assignee,
     projects.assignee_id,
@@ -1860,6 +1866,13 @@ CREATE INDEX index_projects_on_incharge_id ON public.projects USING btree (incha
 
 
 --
+-- Name: index_projects_on_kam_assignee_id; Type: INDEX; Schema: public; Owner: -
+--
+
+CREATE INDEX index_projects_on_kam_assignee_id ON public.projects USING btree (kam_assignee_id);
+
+
+--
 -- Name: index_projects_on_kam_region_id; Type: INDEX; Schema: public; Owner: -
 --
 
@@ -2137,6 +2150,14 @@ ALTER TABLE ONLY public.active_storage_variant_records
 
 
 --
+-- Name: projects fk_rails_993c2a6f6a; Type: FK CONSTRAINT; Schema: public; Owner: -
+--
+
+ALTER TABLE ONLY public.projects
+    ADD CONSTRAINT fk_rails_993c2a6f6a FOREIGN KEY (kam_assignee_id) REFERENCES public.telco_uam_users(id);
+
+
+--
 -- Name: projects fk_rails_99fc2a1a9e; Type: FK CONSTRAINT; Schema: public; Owner: -
 --
 
@@ -2305,6 +2326,12 @@ INSERT INTO "schema_migrations" (version) VALUES
 ('20211220120052'),
 ('20220103082627'),
 ('20220213091922'),
-('20220306212410');
+('20220228080720'),
+('20220306212410'),
+('20220315102300'),
+('20220318172640'),
+('20220322150514'),
+('20220324063336'),
+('20220325092640');