+++ spec/models/project_spec.rb
@@ -31,40 +31,6 @@ RSpec.describe Project, type: :model do
     it { is_expected.to validate_presence_of(:address) }
     it { is_expected.to validate_uniqueness_of(:external_id).allow_nil }
 
-    context 'when move in end date is before the start date' do
-      subject(:project) { build(:project, move_in_starts_on: Date.current, move_in_ends_on: Date.current.yesterday) }
-
-      it 'invalidates record' do
-        expect(project).not_to be_valid
-        expect(project.errors.full_messages).to eq(
-          [
-            "Move in ends on #{t('date.errors.messages.must_be_after',
-                                 date: project.move_in_starts_on)}"
-          ]
-        )
-      end
-    end
-
-    context 'when move in end date equal to the start date' do
-      subject(:project) { build(:project, move_in_starts_on: Date.current, move_in_ends_on: Date.current) }
-
-      it 'invalidates record' do
-        expect(project).not_to be_valid
-        expect(project.errors.full_messages).to eq(
-          [
-            "Move in ends on #{t('date.errors.messages.must_be_after',
-                                 date: project.move_in_starts_on)}"
-          ]
-        )
-      end
-    end
-
-    context 'when move in end date is after the start date' do
-      it do
-        expect(build(:project, move_in_starts_on: Date.current, move_in_ends_on: Date.current.tomorrow)).to be_valid
-      end
-    end
-
     context 'when move in start date is before the construction date' do
       subject(:project) do
         build(:project, construction_starts_on: Date.current, move_in_starts_on: Date.current.yesterday)
@@ -194,6 +160,7 @@ RSpec.describe Project, type: :model do
         technical_analysis: 'Technical Analysis',
         technical_analysis_completed: 'Technical Analysis Completed',
         ready_for_offer: 'Ready for Offer',
+        offer_confirmation: 'Offer Confirmation',
         contract: 'Contract',
         contract_accepted: 'Contract Accepted',
         under_construction: 'Under Construction',
@@ -215,6 +182,15 @@ RSpec.describe Project, type: :model do
         nursing: 'Nursing',
         retirement_homes: 'Retirement Homes',
         others: 'Others'
+      ).backed_by_column_of_type(:string)
+    end
+
+    it do
+      expect(subject).to define_enum_for(:confirmation_status).with_values( # rubocop:disable RSpec/NamedSubject
+        new_offer: 'New',
+        negotiation: 'Negotiation',
+        offered: 'Offered',
+        signed: 'Signed'
       ).backed_by_column_of_type(:string)
     end
   end